Core functions completely implemented
At this commit, all functions have been implemented, but a lot of stuff is
commented out. The library has lots of non-idiomatic code, unsafe and weird
cases. Plus, it contains a fork of the BTreeSet, which is not the best way
to go about things. The methods won't work, but at least the core of the
algorithm is done. You can clearly see what functions are needed, what
functions are called, etc. Before this library is put into testing, the data
structures should be made more idiomatic.

It should be noted that possible_intersection(2) is only called from the
calculate() function. divide_segment(5) is only called from the
possible_intersect(2) function. So it would make sense to inline / nest
these two functions to make their scope more clear.